[Add]homes controller spec
