+++ spec/requests/public/cocktails_request_spec.rb
