+++ spec/requests/public/end_users_request_spec.rb
