+++ spec/requests/public/favorites_request_spec.rb
