+++ spec/requests/public/homes_request_spec.rb
