+++ spec/requests/public/rates_request_spec.rb
